Fixed some bugs related to progress updating and Nldas Forcing
Added checks in ProgressUpdater and fixed issue with new race condition
in ErrorLog.
Modified printout in DatabaseCache.
Added check for existing file to NldasForcing Indices.

+++ src/version2/prototype/indices/NldasForcing/NldasForcingFreezingDegreeDays.java
@@ -26,7 +26,9 @@ public class NldasForcingFreezingDegreeDays extends IndicesFramework {
             if(input.getName().contains("FreezingDegreeDays"))
             {
                 FileUtils.forceMkdir(mOutputFile.getParentFile());
-                Files.copy(input.toPath(), mOutputFile.toPath());
+                if(!mOutputFile.exists()) {
+                    Files.copy(input.toPath(), mOutputFile.toPath());
+                }
                 //                if(!input.renameTo(mOutputFile)) {
                 //                    CopyFile(input, mOutputFile);
                 //                }
